Rename relay client factory to vuex-relay-adapters
Language switch. This thing did not belong in utilities, and it is more
of an adapter between the relay client, and the rest of the Vuejs infra-
structre
